channels deleted when server shuts down, no more leaks
