+++ include/User.hpp
@@ -14,7 +14,6 @@ class	User
 	public:
 		User(int fd, Server* server);
 		~User();
-        User(int fd); // needed? -> not private?
 
 		void				setNickname(const std::string& nickname);
 		void				setUsername(const std::string& username);
