+++ src/Server.cpp
@@ -14,6 +14,7 @@
 #include "../include/signal.hpp"	// g_running variable
 #include "../include/utils.hpp"		// getFormattedTime(), logServerMessage()
 
+/// Constructor: Initializes the server socket and sets up the server state.
 Server::Server(int port, const std::string& password) 
 	:	_name(SERVER_NAME), _version(VERSION), _network(NETWORK),
 		_creationTime(getFormattedTime()), _port(port),
@@ -23,6 +24,7 @@ Server::Server(int port, const std::string& password)
 	initSocket();
 }
 
+// Destructor: Closes the server socket and cleans up resources.
 Server::~Server()
 {
 	// Close the listening socket if open
@@ -32,6 +34,10 @@ Server::~Server()
 	// Delete all dynamically allocated User objects
 	while (!_usersFd.empty())
 		deleteUser(_usersFd.begin()->first);
+
+	// Delete all dynamically allocated Channel objects
+	while (!_channels.empty())
+		deleteChannel(_channels.begin()->first);
 
 	logServerMessage("Server shutdown complete");
 }
@@ -79,6 +85,32 @@ void	Server::run()
 		// Handle user output for all users with pending data
 		handleWriteReadyUsers(writeFds);
 	}
+}
+
+/**
+ Handles a failed `send()` operation to a user.
+
+ Logs the error to the server terminal.
+ If the failure was due to a broken or reset connection (`EPIPE` or `ECONNRESET`),
+ the user is considered disconnected and is removed from the server.
+
+ @param fd 		The fd of the user for whom `send()` failed.
+ @param nick 	The nickname of the user for whom `send()` failed.
+*/
+void	Server::handleSendError(int fd, const std::string& nick)
+{
+	// Critical errors that mean the user is gone
+	if (errno == EPIPE || errno == ECONNRESET)
+	{
+		handleDisconnection(fd, strerror(errno), "send()");
+		deleteUser(fd);
+		return;
+	}
+
+	// Non‑critical send error (temporary): log it but do not disconnect
+	// e.g. EAGAIN or EWOULDBLOCK (“Resource temporarily unavailable”)
+	logUserAction(nick, fd, RED + std::string("Error sending message to user: ")
+		+ strerror(errno) + RESET);
 }
 
 /////////////
@@ -133,6 +165,10 @@ int	Server::getMaxChannels() const
 	return _maxChannels;
 }
 
+//////////////////
+// Nick Mapping //
+//////////////////
+
 // Returns a map of active users by nickname.
 std::map<std::string, User*>&	Server::getNickMap()
 {
@@ -144,65 +180,4 @@ std::map<std::string, User*>&	Server::getNickMap()
 void	Server::removeNickMapping(const std::string& nickname)
 {
 	_usersNick.erase(nickname);
-}
-
-void	Server::addChannel(Channel* channel)
-{
-	if (channel)
-		_channels[channel->get_name()] = channel;
-}
-
-/**
- Retrieves an `Channel` object by its name.
-
- @param channelName 	The name of the channel to retrieve.
- @return				Pointer to the `Channel` object if found, `NULL` otherwise.
-*/
-Channel*	Server::getChannel(const std::string& channelName) const
-{
-	std::map<std::string, Channel*>::const_iterator	it = _channels.find(channelName);
-	if (it != _channels.end())
-		return it->second;
-	return NULL;
-}
-
-/**
- Retrieves an existing `Channel` by name or creates a new one if it does not exist.
-
- If the channel with the given name exists, a pointer to it is returned.
- Otherwise, a new `Channel` object is created, added to the server, and returned.
- In case of memory allocation failure, an error is logged and the user is notified.
-
- @param channelName 	The name of the channel to retrieve or create.
- @param user 			The user requesting or triggering the channel creation.
- @param key 			Optional channel key to set if creating a new channel.
- @return				Pointer to the `Channel` object, or `NULL` on failure.
-*/
-Channel*	Server::getOrCreateChannel(const std::string& channelName, User* user, const std::string& key)
-{
-	Channel*	channel = getChannel(channelName);
-	if (channel)
-		return channel;
-
-	// Create a new channel if it does not exist
-	try
-	{
-		channel = new Channel(channelName);
-		if (!key.empty()) // If a key is provided, set it as the channel password
-			channel->set_password(key);
-		addChannel(channel);
-
-		logUserAction(user->getNickname(), user->getFd(),
-			std::string("created and joined new channel: ") + BLUE + channelName + RESET);
-	}
-	catch(const std::bad_alloc&)
-	{
-		logServerMessage(RED + std::string("ERROR: Failed to allocate memory for new channel ")
-			+ BLUE + channelName + RED + " (created by " + GREEN + user->getNickname() 
-			+ RED + ", " + MAGENTA + "fd " + toString(user->getFd()) + RED + ")" + RESET);
-
-		user->replyError(500, "", "Internal server error while creating channel " + channelName);
-		return NULL;
-	}
-	return channel;
 }
